+++ packages/frontend/src/app/events/list/list.component.ts
@@ -9,13 +9,19 @@ import { EventService } from 'src/app/core/services/event.service'
 })
 export class ListComponent implements OnInit {
   public events: Event[] = []
+  public selectedEvent!: Event | null
 
   constructor(private _eventService: EventService) {}
 
   ngOnInit(): void {
-    this._eventService.getEvents().subscribe(events => {
-      console.log(events)
-      this.events = events
-    })
+    this._eventService.getEvents().subscribe(events => (this.events = events))
+
+    this._eventService.$selectedEvent.subscribe(
+      event => (this.selectedEvent = event)
+    )
+  }
+
+  public handleOpen(event: Event) {
+    this._eventService.selectEvent(event)
   }
 }