+++ packages/cms/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/Miwoli/jugger/compare/cms@0.5.0...cms@0.6.0) (2022-12-13)
+
+### Features
+
+- wIP: Event display improvement ([8e4f7da](https://github.com/Miwoli/jugger/commit/8e4f7daff020546774ceeb0056a70336ae9f4916))
+
 # [0.5.0](https://github.com/Miwoli/jugger/compare/cms@0.4.2...cms@0.5.0) (2022-12-07)
 
 ### Features